Setup: separated class to read configs
